requirements.txt
llama-index
# llama-index-llms-ollama
llama-index-llms-openai
llama-index-embeddings-ollama
llama-index-llms-groq
python-dotenv
